Move tresult into sktm.misc
Move sktm.tresult to sktm.misc.tresult to fix composition, so that
modules lower in the hierarchy (e.g. sktm.jenkins) do not have to import
the module above (i.e. sktm).

Fixes #6
