+++ sktm/db.py
@@ -17,7 +17,7 @@ import logging
 import os
 import sqlite3
 import time
-import sktm
+import sktm.misc
 
 
 class skt_db(object):
@@ -350,7 +350,7 @@ class skt_db(object):
         if not result:
             return None
 
-        return sktm.tresult(result[0])
+        return sktm.misc.tresult(result[0])
 
     def get_stable(self, baserepo):
         """Get the latest stable commit ID for a baseline Git repo URL.
@@ -631,7 +631,7 @@ class skt_db(object):
         for (burl, commit, res, buildid) in self.cur.fetchall():
             print("repo url:", burl)
             print("commit id:", commit)
-            print("result:", sktm.tresult(res).name)
+            print("result:", sktm.misc.tresult(res).name)
             print("build id: #", buildid, sep='')
             print("---")
