+++ sktm/misc.py
@@ -1,0 +1,25 @@
+# Copyright (c) 2017-2018 Red Hat, Inc. All rights reserved. This copyrighted
+# material is made available to anyone wishing to use, modify, copy, or
+# redistribute it subject to the terms and conditions of the GNU General
+# Public License v.2 or later.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software Foundation, Inc.,
+# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+
+import enum
+
+
+class tresult(enum.IntEnum):
+    """Test result"""
+    SUCCESS = 0
+    MERGE_FAILURE = 1
+    BUILD_FAILURE = 2
+    PUBLISH_FAILURE = 3
+    TEST_FAILURE = 4
+    BASELINE_FAILURE = 5
